Fix vuex templates
