--- frontend/src/store/getters.js
@@ -1,11 +1,0 @@
-export default {
-    getUser(state) {
-        return state.user;
-    },
-    getChatState(state){
-        return state.ChatState
-    },
-    getCalendarRecruit(state){
-        return state.calendarRecruit
-    }
-}
