+++ frontend/src/store/index.js
@@ -1,18 +1,18 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import getters from './getters'
-import actions from './actions'
-import mutations from './mutations'
-import state from './state'
 import axios from 'axios'
+import authentication from './authentication'
+import recruit from './recruit'
+import resume from './resume'
+import template from './template'
 
 Vue.use(Vuex, axios)
 
 export default new Vuex.Store({
-  state,
-  getters,
-  mutations,
-  actions,
-  // modules: {
-  // }
+  modules: {
+    auth : authentication,
+    recruit : recruit,
+    resume : resume,
+    template : template,
+  }
 })
