+++ frontend/src/store/resume.js
@@ -1,0 +1,58 @@
+export default {
+  state: {
+    resume_list: {
+      1: {
+        1: [
+          {
+            "companyId": 1,
+            "recruitId": 13,
+            "companyName": "스마일 게이트",
+            "startTime": "2019-12-29 15:00:00",
+            "endTime": "2019-12-30 15:00:00",
+            "recruitType": 1,
+            "employType": [
+              1,
+              3
+            ]
+          },
+          {
+            "companyId": 1,
+            "recruitId": 12,
+            "companyName": "네이버",
+            "startTime": "2019-12-29 15:00:00",
+            "endTime": "2020-01-22 10:00:00",
+            "recruitType": 1,
+            "employType": [
+              1,
+              3
+            ]
+          },
+        ],
+        2: [],
+        3: [{
+          "companyId": 1,
+          "recruitId": 22,
+          "companyName": "카카오",
+          "startTime": "2019-12-29 15:00:00",
+          "endTime": "2020-01-28 10:00:00",
+          "recruitType": 1,
+          "employType": [
+            1,
+            3
+          ]
+        },
+        ],
+      },
+      2: { 1: [], 2: [] },
+      3: { 1: [], 2: [] },
+      4: { 1: [], 2: [] },
+      5: { 1: [], 2: [] },
+
+
+
+    }
+  },
+  mutations: {},
+  actions: {},
+  getters: {}
+}
