--- frontend/src/store/state.js
@@ -1,313 +1,0 @@
-export default {
-  user :{
-    email : "",
-    nickname : ""
-  },
-  ChatState: false, // 채팅방 열림 여부
-  starRank: {
-    1: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    2: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    3: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    4: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    5: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-  },
-  visitedRank: {
-    1: {
-      company: "스마일게이트2",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    2: {
-      company: "스마일게이트2",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    3: {
-      company: "스마일게이트2",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    4: {
-      company: "스마일게이트2",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    5: {
-      company: "스마일게이트2",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-  },
-  ApplicantRank: {
-    1: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    2: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    3: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    4: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-    5: {
-      company: "스마일게이트",
-      remainingDays: 5,
-      count: 500,
-      companyId: '111222'
-    },
-  },
-  calendarRecruit: [
-    {
-      "companyId": 1,
-      "recruitId": 1,
-      "companyName": "회사1이무니다",
-      "startTime": "2019-12-29 15:00:00",
-      "endTime": "2019-12-30 15:00:00",
-      "recruitType": 1,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        1,
-        3
-      ]
-    },
-    {
-      "companyId": 1,
-      "recruitId": 2,
-      "companyName": "회사1",
-      "startTime": "2019-12-30 15:00:00",
-      "endTime": "2020-02-01 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        1,
-        1
-      ]
-    },
-    {
-      "companyId": 1,
-      "recruitId": 3,
-      "companyName": "회사1",
-      "startTime": "2019-12-31 15:00:00",
-      "endTime": "2020-01-05 15:00:00",
-      "recruitType": 1,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        3
-      ]
-    },
-    {
-      "companyId": 2,
-      "recruitId": 4,
-      "companyName": "회사2",
-      "startTime": "2019-12-29 15:00:00",
-      "endTime": "2020-01-25 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        1
-      ]
-    },
-    {
-      "companyId": 5,
-      "recruitId": 5,
-      "companyName": "회사5",
-      "startTime": "2019-12-29 15:00:00",
-      "endTime": "2020-01-08 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        1
-      ]
-    },
-    {
-      "companyId": 2,
-      "recruitId": 6,
-      "companyName": "회사2",
-      "startTime": "2019-12-30 15:00:00",
-      "endTime": "2020-02-01 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        1
-      ]
-    },
-    {
-      "companyId": 3,
-      "recruitId": 7,
-      "companyName": "회사3",
-      "startTime": "2020-01-10 15:00:00",
-      "endTime": "2020-02-18 15:00:00",
-      "recruitType": 1,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        1
-      ]
-    },
-    {
-      "companyId": 2,
-      "recruitId": 8,
-      "companyName": "회사2",
-      "startTime": "2020-12-29 15:00:00",
-      "endTime": "2020-02-01 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        2
-      ]
-    },
-    {
-      "companyId": 2,
-      "recruitId": 9,
-      "companyName": "회사2",
-      "startTime": "2019-12-30 15:00:00",
-      "endTime": "2020-02-01 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        3
-      ]
-    },
-    {
-      "companyId": 3,
-      "recruitId": 10,
-      "companyName": "회사3",
-      "startTime": "2019-01-13 15:00:00",
-      "endTime": "2020-02-18 15:00:00",
-      "recruitType": 1,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        2
-      ]
-    },
-    {
-      "companyId": 3,
-      "recruitId": 11,
-      "companyName": "회사3",
-      "startTime": "2019-12-29 15:00:00",
-      "endTime": "2020-02-01 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        2
-      ]
-    },
-    {
-      "companyId": 3,
-      "recruitId": 12,
-      "companyName": "회사3",
-      "startTime": "2020-01-03 15:00:00",
-      "endTime": "2020-01-31 15:00:00",
-      "recruitType": 0,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        2
-      ]
-    },
-    {
-      "companyId": 4,
-      "recruitId": 13,
-      "companyName": "회사4",
-      "startTime": "2020-01-03 15:00:00",
-      "endTime": "2020-02-28 15:00:00",
-      "recruitType": 1,
-      "imageFileName": "https://jasoseol.s3-ap-northeast-2.amazonaws.com/employment_companies/images/000/035/504/original/AXA_Logo.png?1577774955",
-      "employType": [
-        3
-      ]
-    }
-  ],
-  resume_list: {
-    1: {
-      1: [
-        {
-          "companyId": 1,
-          "recruitId": 13,
-          "companyName": "스마일 게이트",
-          "startTime": "2019-12-29 15:00:00",
-          "endTime": "2019-12-30 15:00:00",
-          "recruitType": 1,
-          "employType": [
-            1,
-            3
-          ]
-        },
-        {
-          "companyId": 1,
-          "recruitId": 12,
-          "companyName": "네이버",
-          "startTime": "2019-12-29 15:00:00",
-          "endTime": "2020-01-22 10:00:00",
-          "recruitType": 1,
-          "employType": [
-            1,
-            3
-          ]
-        },
-      ],
-      2: [],
-      3: [{
-        "companyId": 1,
-        "recruitId": 22,
-        "companyName": "카카오",
-        "startTime": "2019-12-29 15:00:00",
-        "endTime": "2020-01-28 10:00:00",
-        "recruitType": 1,
-        "employType": [
-          1,
-          3
-        ]},
-      ],
-    },
-    2:{1:[],2:[]},
-    3:{1:[],2:[]},
-    4:{1:[],2:[]},
-    5:{1:[],2:[]},
-
-
-
-  }
-}
